migrate units from last profile
